fix(mobile): Swift 6 concurrency + CoreML linker + Compose fixes
- Swift 6 strict concurrency: replaced vm_kernel_page_size (global mutable C var)
  with hardcoded ARM64 page size in PantheonApp + WidgetIntents
- BrainView: replaced NSDirectoryEnumerator with FileManager.subpaths in
  Task.detached (NSDirectoryEnumerator not Sendable)
- PantheonBridge: removed Int64 casts (gomobile int maps to Swift Int)
- project.yml: added CoreML.framework to main Pantheon target linker flags
- SteleScreen: CircularShape → CircleShape (correct Compose API)
- iOS BUILD SUCCEEDED, Android assembleRelease BUILD SUCCESSFUL

+++ ios/Pantheon/Services/PantheonBridge.swift
@@ -164,14 +164,14 @@ final class PantheonBridge: Sendable {
 
     func vaultStore(source: String, tag: String, content: String, tokens: Int) async throws -> VaultEntry {
         return try await Task.detached(priority: .userInitiated) {
-            let json = MobileVaultStore(source, tag, content, Int64(tokens))
+            let json = MobileVaultStore(source, tag, content, tokens)
             return try self.decode(json) as VaultEntry
         }.value
     }
 
     func vaultSearch(query: String, limit: Int = 10) async throws -> VaultSearchResult {
         return try await Task.detached(priority: .userInitiated) {
-            let json = MobileVaultSearch(query, Int64(limit))
+            let json = MobileVaultSearch(query, limit)
             return try self.decode(json) as VaultSearchResult
         }.value
     }
@@ -192,7 +192,7 @@ final class PantheonBridge: Sendable {
 
     func vaultPrune(olderThanHours: Int) async throws -> VaultPruneResult {
         return try await Task.detached(priority: .userInitiated) {
-            let json = MobileVaultPrune(Int64(olderThanHours))
+            let json = MobileVaultPrune(olderThanHours)
             return try self.decode(json) as VaultPruneResult
         }.value
     }
@@ -231,7 +231,7 @@ final class PantheonBridge: Sendable {
 
     func steleReadRecent(count: Int) async throws -> [SteleEntry] {
         return try await Task.detached(priority: .userInitiated) {
-            let json = MobileSteleReadRecent(Int64(count))
+            let json = MobileSteleReadRecent(count)
             return try self.decode(json) as [SteleEntry]
         }.value
     }
